(feat) Lots of code restructuring, add some response helpers
